requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
boto3==1.34.120
pypdf==4.2.0
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.35
langchain-chroma==0.2.6
chromadb==1.5.8
tiktoken==0.8.0
marshmallow==3.21.3
